pyproject.toml
[project]
name = "geminiclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docker>=7.1.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "psycopg[binary,pool]>=3.2",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
google = [
    "google-adk>=1.27.2",
    "google-genai>=1.3.0",
]
deep_search = [
    "fastembed>=0.7.4",
    "qdrant-client>=1.17.1",
]

[project.scripts]
geminiclaw = "src.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests (require real API)",
]

